+++ gradle/libs.versions.toml
@@ -4,7 +4,6 @@ kotlin = "2.0.0"
 ksp = "2.0.0-1.0.24" # Must be compatible with: `kotlin`
 desugarLibs = "1.1.5"
 androidxLifecycle = "2.8.6"
-composeCompiler = "1.5.14" # Must be compatible with: `kotlin`
 androidxComposeBom = "2024.09.00"
 androidxActivity = "1.9.2"
 decompose = "3.1.0"
