+++ shared/arkitekt-decompose/build.gradle.kts
@@ -5,6 +5,11 @@ plugins {
     id(libs.plugins.com.android.library.get().pluginId)
     id(libs.plugins.kotlin.multiplatform.get().pluginId)
     id(libs.plugins.conventions.lint.get().pluginId)
+    alias(libs.plugins.compose.compiler)
+}
+
+dependencies {
+    implementation(platform(libs.androidx.compose.bom))
 }
 
 kotlin {
@@ -19,6 +24,11 @@ kotlin {
     iosTargets()
 
     sourceSets {
+        androidMain {
+            dependencies {
+                implementation(libs.androidx.compose.runtime)
+            }
+        }
         commonMain {
             dependencies {
                 implementation(libs.decompose)
